+++ Frontend/src/lib/components/icons/checkmark.svelte
@@ -2,6 +2,6 @@
   export let className: string | undefined = '!w-6 !h-6 !fill-green-500';
 </script>
 
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" id="small-icon" class={className}>
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" class={className + ' small-icon'}>
   <path d="M0 11l2-2 5 5L18 3l2 2L7 18z" />
 </svg>