+++ CHANGELOG.md
@@ -5,12 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.13] - 2024-10-15
+
+### Added
+
+- Add Capabilities Section
+- Add tab event to scroll sections
+- Add question to pricing section
+
 ## [0.0.12] - 2024-10-14
 
 ### Added
 
 - Add products section on homepage
-- Add tab event to scroll
+- Add dashboard image
 
 ### Changed